server/requirements.txt
asgiref==3.4.1
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.4
decorator==5.0.7
Django==3.2.6
django-bootstrap-v5==1.0.2
django-cors-headers==3.13.0
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
executing==1.0.0
Faker==8.1.1
idna==3.4
importlib-metadata==2.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.2.0
matplotlib-inline==0.1.3
numpy==1.23.5
pandas==1.5.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pure-eval==0.2.2
Pygments==2.8.1
PyJWT==2.6.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.15.0
sklearn==0.0.post1
soupsieve==2.2.1
sqlparse==0.4.1
stack-data==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
traitlets==5.0.5
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.4.1
